requirements.txt
browser-use>=0.12.0
groq>=0.9.0
langchain-openai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.2.0
weasyprint>=60.0
aiofiles>=23.0.0
